requirements.txt
edge-tts>=6.1.0
python-dotenv>=1.0.0
requests>=2.31.0
gradio>=5.0.0
